REF: Placed CoolDownTracker into its own file ENH: More comments REF: Extracted the building of custom CoolDownTrackers according to a config file into its own method ADD: If a WHOIS server is not found, an empty response is given.
